pyproject.toml
[project]
name = "matchcut-generator"
version = "0.1.0"
description = "makes all the matchcut crops automatically"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "setuptools>=80.10.2",
    "undetected-chromedriver>=3.5.5",
    "pytesseract>=0.3.13",
    "Pillow>=11.1.0",
    "numpy>=2.4.1",
    "ddgs>=9.10.0",
    "opencv-contrib-python>=4.13.0.90",
]
